Rifactoring di Vector e position

+++ src/main/java/edu/wol/server/connector/ws/encoders/PositionSerializer.java
@@ -16,9 +16,9 @@ public class PositionSerializer implements JsonSerializer<Position> {
 	public JsonElement serialize(Position src, Type typeOfSrc,
 			JsonSerializationContext context) {
 		JsonObject jsonPos=new JsonObject();
-		jsonPos.add("x", new JsonPrimitive(new Long(src.getX())));
-		jsonPos.add("y", new JsonPrimitive(new Long(src.getY())));
-		jsonPos.add("z", new JsonPrimitive(new Long(src.getZ())));
+		jsonPos.add("x", new JsonPrimitive(new Float(src.getX())));
+		jsonPos.add("y", new JsonPrimitive(new Float(src.getY())));
+		jsonPos.add("z", new JsonPrimitive(new Float(src.getZ())));
 		return jsonPos;
 	}
